requirements.txt
# Claude Agent Requirements
# Python 3.12 recommended

# Core
requests>=2.28.0

# Audio
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
pydub>=0.25.1

# TTS (choose one or more)
edge-tts>=6.1.0
# chatterbox-tts  # Uncomment for Chatterbox (needs GPU)
# piper-tts       # Uncomment for Piper

# Vision (optional)
opencv-python>=4.8.0
Pillow>=10.0.0

# For vision descriptions, install Ollama separately:
# https://ollama.ai
# Then run: ollama pull llava:7b

# Note: numpy version matters for some audio libs
numpy<2.3
